+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 
 ## [Unreleased]
+### Added
+- Added races to Valhalla: Dwarves, Goblins, Firewalkers, Lizardmen
 
 ## [0.6.0-2] - 2019-04-09
 ### Changed